translate analyzer
